clearing up old implementations
